Do not use my_make_scrambled_password()
my_make_scrambled_password() is not meant to be a replacement for
the removed make_scrambled_password() function from libmysqlclient.

So look for make_scrambled_password() in the mysql client library
instead of my_make_scrambled_password(). If not found, then
reimplement it, but with the correct name to avoid further confusion
around this.